requirements.txt
virtualenv==16.2.0
Fabric3==1.14.post1
autopep8==1.4.3
pycodestyle==2.4.0
pytest==4.1.1
botocore==1.12.83
boto3==1.9.83
awscli==1.16.93
slackclient==1.3.0
wheel